+++ src/main/java/edu/harvard/iq/dataverse/authorization/providers/oauth2/oidc/OIDCAuthProvider.java
@@ -101,20 +101,6 @@ public class OIDCAuthProvider extends AbstractOAuth2AuthenticationProvider {
         this.pkceMethod = CodeChallengeMethod.parse(pkceMethod);
     }
     
-    /**
-     * Although this is defined in {@link edu.harvard.iq.dataverse.authorization.AuthenticationProvider},
-     * this needs to be present due to bugs in ELResolver (has been modified for Spring).
-     * TODO: for the future it might be interesting to make this configurable via the provider JSON (it's used for ORCID!)
-     * @see <a href="https://issues.jboss.org/browse/JBEE-159">JBoss Issue 159</a>
-     * @see <a href="https://github.com/eclipse-ee4j/el-ri/issues/43">Jakarta EE Bug 43</a>
-     * @return false
-     */
-    
-    //ToDo: Should be fixed - don't need this now
-    @Override
-    public boolean isDisplayIdentifier() {
-        return false;
-    }
     
     /**
      * Setup metadata from OIDC provider during creation of the provider representation
@@ -273,7 +259,7 @@ public class OIDCAuthProvider extends AbstractOAuth2AuthenticationProvider {
      * @param userInfo
      * @return the usable user record for processing ing {@link edu.harvard.iq.dataverse.authorization.providers.oauth2.OAuth2LoginBackingBean}
      */
-    OAuth2UserRecord getUserRecord(UserInfo userInfo, int consentVersion) {
+    public OAuth2UserRecord getUserRecord(UserInfo userInfo, int consentVersion) {
         boolean usesMFA = false;
         String role = userInfo.getStringClaim("role");
         role = role == null ? "" : role;
@@ -334,7 +320,7 @@ public class OIDCAuthProvider extends AbstractOAuth2AuthenticationProvider {
      * Retrieve User Info from provider. Encapsulate for testing.
      * @param accessToken The access token to enable reading data from userinfo endpoint
      */
-    Optional<UserInfo> getUserInfo(BearerAccessToken accessToken) throws IOException, OAuth2Exception {
+    public Optional<UserInfo> getUserInfo(BearerAccessToken accessToken) throws IOException, OAuth2Exception {
         // Retrieve data
         HTTPResponse response = new UserInfoRequest(this.idpMetadata.getUserInfoEndpointURI(), accessToken)
                                         .toHTTPRequest()
@@ -362,50 +348,4 @@ public class OIDCAuthProvider extends AbstractOAuth2AuthenticationProvider {
         }
     }
 
-    /**
-     * Trades an access token for an {@link UserRecordIdentifier} (if valid).
-     *
-     * @apiNote The resulting {@link UserRecordIdentifier} may be used with
-     *          {@link edu.harvard.iq.dataverse.authorization.AuthenticationServiceBean#lookupUser(UserRecordIdentifier)}
-     *          to look up an {@link edu.harvard.iq.dataverse.authorization.users.AuthenticatedUser} from the database.
-     * @see edu.harvard.iq.dataverse.api.auth.BearerTokenAuthMechanism
-     *
-     * @param accessToken The token to use when requesting user information from the provider
-     * @return Returns an {@link UserRecordIdentifier} for a valid access token or an empty {@link Optional}.
-     * @throws IOException In case communication with the endpoint fails to succeed for an I/O reason
-     */
-    public Optional<UserRecordIdentifier> getUserIdentifier(BearerAccessToken accessToken) throws IOException {
-        OAuth2UserRecord userRecord;
-        try {
-            // Try to retrieve with given token (throws if invalid token)
-            Optional<UserInfo> userInfo = getUserInfo(accessToken);
-            
-            if (userInfo.isPresent()) {
-                // Take this detour to avoid code duplication and potentially hard to track conversion errors.
-                int consentVersion = 0;
-                try {
-                    JWTClaimsSet claims = JWTClaimsSet.parse(accessToken.toJSONObject());
-                    consentVersion = Integer.parseInt(claims.getStringClaim("consentver"));
-                } catch (java.text.ParseException | NumberFormatException e) {
-                    logger.warning("Could not parse access token: " + accessToken.toJSONString() + ", err: " + e.getLocalizedMessage());
-                }
-                userRecord = getUserRecord(userInfo.get(), consentVersion);
-            } else {
-                // This should not happen - an error at the provider side will lead to an exception.
-                logger.log(Level.WARNING,
-                    "User info retrieval from {0} returned empty optional but expected exception for token {1}.",
-                    List.of(getId(), accessToken).toArray()
-                );
-                return Optional.empty();
-            }
-        } catch (OAuth2Exception e) {
-            logger.log(Level.FINE,
-                "Could not retrieve user info with token {0} at provider {1}: {2}",
-                List.of(accessToken, getId(), e.getMessage()).toArray());
-            logger.log(Level.FINER, "Retrieval failed, details as follows: ", e);
-            return Optional.empty();
-        }
-        
-        return Optional.of(userRecord.getUserRecordIdentifier());
-    }
 }