+++ src/main/java/edu/harvard/iq/dataverse/search/SearchFields.java
@@ -171,6 +171,7 @@ public class SearchFields {
     public static final String FILE_CHECKSUM_TYPE = "fileChecksumType";
     public static final String FILE_CHECKSUM_VALUE = "fileChecksumValue";
     public static final String FILENAME_WITHOUT_EXTENSION = "fileNameWithoutExtension";
+    public static final String FILE_RESTRICTED = "fileRestricted";
     /**
      * Indexed as a string so we can facet on it.
      */
@@ -233,7 +234,8 @@ public class SearchFields {
      */
     public static final String PUBLICATION_STATUS = "publicationStatus";
     
-    public static final String EXTERNAL_STATUS = "externalStatus";
+    public static final String CURATION_STATUS = "curationStatus";
+    public static final String CURATION_STATUS_CREATE_TIME = "curationStatusCreateTime";
     /**
      * @todo reconcile different with Solr schema.xml where type is Long rather
      * than String.
@@ -271,6 +273,8 @@ public class SearchFields {
      */
     public static final String DATASET_TYPE = "datasetType";
 
+    public static final String OBSERVATIONS = "observations";
+    public static final String VARIABLE_COUNT = "variableCount";
     public static final String VARIABLE_NAME = "variableName";
     public static final String VARIABLE_LABEL = "variableLabel";
     public static final String LITERAL_QUESTION = "literalQuestion";